requirements.txt
astroid==2.3.3
autopep8==1.5.1
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
docxtpl==0.6.3
Flask==1.1.1
html5lib==1.0.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
mechanize==0.4.5
numpy==1.18.4
pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
PyPDF2==1.26.0
python-docx==0.8.7
python-dotenv==0.14.0
pytube==9.6.0
pytube3==9.6.4
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
typing-extensions==3.7.4.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.11.2
